+++ tests/testthat/test-filter_trip_id.R
@@ -53,9 +53,9 @@ test_that("doesn't change given gtfs", {
 
 test_that("'trip_id' and 'keep' arguments work correctly", {
   smaller_keeping <- filter_trip_id(ggl_gtfs, ggl_trips)
-  expect_true(any(smaller_keeping$trips$trip_id %chin% ggl_trips))
-  expect_true(any(smaller_keeping$stop_times$trip_id %chin% ggl_trips))
-  expect_true(any(smaller_keeping$frequencies$trip_id %chin% ggl_trips))
+  expect_true(all(smaller_keeping$trips$trip_id %chin% ggl_trips))
+  expect_true(all(smaller_keeping$stop_times$trip_id %chin% ggl_trips))
+  expect_true(all(smaller_keeping$frequencies$trip_id %chin% ggl_trips))
 
   smaller_not_keeping <- filter_trip_id(ggl_gtfs, ggl_trips, keep = FALSE)
   expect_true(!any(smaller_not_keeping$trips$trip_id %chin% ggl_trips))
@@ -107,7 +107,8 @@ test_that("the function filters berlin's gtfs correctly", {
   relevant_stops <- unique(
     ber_gtfs$stop_times[trip_id %chin% ber_trips]$stop_id
   )
-  expect_true(all(smaller_ber$stop_times$stop_id %chin% relevant_stops))
+  relevant_stops <- get_parent_station(ber_gtfs, relevant_stops)$stop_id
+  expect_true(all(smaller_ber$stops$stop_id %chin% relevant_stops))
 })
 
 test_that("the function filters sao paulo's gtfs correctly", {
@@ -142,7 +143,7 @@ test_that("the function filters sao paulo's gtfs correctly", {
   relevant_stops <- unique(
     spo_gtfs$stop_times[trip_id %chin% spo_trips]$stop_id
   )
-  expect_true(all(smaller_spo$stop_times$stop_id %chin% relevant_stops))
+  expect_true(all(smaller_spo$stops$stop_id %chin% relevant_stops))
 
   # frequencies
   expect_true(all(smaller_spo$frequencies$trip_id %chin% spo_trips))